syntropy-mcp/.ce/tools/pyproject.toml
[project]
name = "ce-tools"
version = "0.1.0"
description = "Context Engineering CLI Tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "diagrams>=0.24.4",
    "jsonschema>=4.25.1",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
ce = "ce.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["ce"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
